refactor(core): remove mini_printf header
[no changelog]
