+++ core/embed/rtl/printf.c
@@ -52,7 +52,7 @@
 #include "printf_config.h"
 //#endif
 
-#include <printf/printf.h>
+#include <rtl/printf.h>
 
 #ifdef __cplusplus
 #include <cstdint>
