+++ core/embed/rtl/printf_config.h
